Fixed service classes. Added Exception class
