+++ class-book-impl/src/main/java/com/eschool/classbook/exception/ClassBookException.java
@@ -1,0 +1,7 @@
+package com.eschool.classbook.exception;
+
+public class ClassBookException extends RuntimeException{
+    public ClassBookException(String message) {
+        super(message);
+    }
+}
