+++ class-book-impl/src/main/java/com/eschool/classbook/group/GroupServiceImpl.java
@@ -1,11 +1,11 @@
 package com.eschool.classbook.group;
 
+import com.eschool.classbook.exception.ClassBookException;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import java.awt.print.Pageable;
 import java.time.LocalDateTime;
-import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service
@@ -19,19 +19,19 @@ public class GroupServiceImpl implements GroupService{
 
     @Override
     public Page<GroupEntity> findAll(Pageable pageable) {
-        return (Page<GroupEntity>) groupRepository.findAll();
+        return groupRepository.findAll(pageable);
     }
 
     @Override
     public Optional<GroupEntity> findById(Long id) {
         return Optional.ofNullable(Optional.of(groupRepository.getById(id))
-                .orElseThrow(() -> new NoSuchElementException(String.format("Group with id %d wasn't found", id))));
+                .orElseThrow(() -> new ClassBookException(String.format("Group with id %d wasn't found", id))));
     }
 
     @Override
     public void deleteById(Long id) {
         GroupEntity group = Optional.of(groupRepository.getById(id))
-                .orElseThrow(() -> new NoSuchElementException(String.format("Group with id %d wasn't found", id)));
+                .orElseThrow(() -> new ClassBookException(String.format("Group with id %d wasn't found", id)));
         group.setModifyingDate(LocalDateTime.now());
         group.setDeleted(true);
         groupRepository.save(group);
@@ -46,7 +46,7 @@ public class GroupServiceImpl implements GroupService{
     @Override
     public GroupEntity update(Long id, GroupEntity groupEntity) {
         Optional.of(groupRepository.getById(id))
-                .orElseThrow(() -> new NoSuchElementException(String.format("Group with id %d wasn't found", id)));
+                .orElseThrow(() -> new ClassBookException(String.format("Group with id %d wasn't found", id)));
         groupEntity.setModifyingDate(LocalDateTime.now());
         return groupRepository.save(groupEntity);
     }
