+++ class-book-impl/src/main/java/com/eschool/classbook/scorepage/ScoreRepository.java
@@ -1,9 +1,8 @@
 package com.eschool.classbook.scorepage;
 
-import com.eschool.classbook.group.GroupEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ScorepageRepository extends JpaRepository<ScoreEntity, Long> {
+public interface ScoreRepository extends JpaRepository<ScoreEntity, Long> {
 }
